add new package lib + add try/catch to raise proper exception
